# openapi/openapi.yaml
swagger: '2.0'
info:
    title: 'AdGuard Home'
    description: 'AdGuard Home REST API. Admin web interface is built on top of this REST API.'
    version: '0.101'
schemes:
    - http
basePath: /control
produces:
    - application/json
...
paths:
...
    /filtering/config:
        post:
            tags:
                - filtering
            operationId: filteringConfig
            summary: 'Set filtering parameters'
            parameters:
            - in: "body"
              name: "body"
              required: true
              schema:
                $ref: "#/definitions/FilterConfig"
            responses:
                200:
                    description: OK
...
definitions:
...
    FilterConfig:
        type: "object"
        description: "Filtering settings"
        properties:
            enabled:
                type: "boolean"
            interval:
                type: "integer"
...